fix: can't use modular-analytics without vue package
