+++ src/index.js
@@ -1,5 +1,4 @@
 // @ts-check
-import { inject } from 'vue'; // eslint-disable-line import/no-unresolved
 import analyticModules from './modules';
 
 const globalProperties = {};
@@ -67,27 +66,3 @@ export const Analytics = {
     }
   },
 };
-
-export const VueAnalytics = {
-  install: (Vue) => {
-    if (Vue.version.split('.').at(0) === '2') {
-      Vue.$analytics = Analytics; // eslint-disable-line no-param-reassign
-      Object.defineProperties(Vue.prototype, {
-        $analytics: {
-          get() {
-            return Vue.$analytics;
-          },
-        },
-      });
-    } else {
-      Vue.config.globalProperties.$analytics = Analytics; // eslint-disable-line no-param-reassign
-      Vue.provide('$analytics', Analytics);
-    }
-  },
-};
-
-export function useAnalytics() {
-  return inject('$analytics');
-}
-
-export default VueAnalytics;
